Refactor SemesterCalendar into PlannedSemesterPeriod
