+++ spec/models/parallel_spec.rb
@@ -2,7 +2,6 @@ require 'spec_helper'
 require 'models/parallel'
 require 'models/person'
 require 'sirius/time_converter'
-require 'sirius/semester_calendar'
 
 describe Parallel do
